requirements.txt
SQLAlchemy~=2.0.30
alembic~=1.13.1
flask~=3.0.3
pydantic~=2.7.1
pydantic-settings~=2.2.1
python-dotenv~=1.0.1
flask-restx~=1.3.0
Werkzeug==3.0.1
flasgger~=0.9.7.1
WTForms~=3.1.2
aiosqlite~=0.20.0
pytest~=8.2.1
httpx~=0.27.0
asyncmy~=0.2.1